Use alternative to NmeaListener
NmeaListener is not always useful, since most of the GGA sentences that
it provides are empty. So to cope up with that, we are using normal
location listener of android and converting its location to a GGA
sentence manually. Now, GGA sentences from NmeaListener and location
listener are being used in a hybrid mode with priority to NmeaListener